refactor: cut the bloat left by the five parallel tracks
Five feature tracks landed in parallel over two days and left duplication
behind. This is the cleanup pass: less code, cleaner seams, no behavior change.
No persisted localStorage key or id format changes.

- presets/labViews: extract src/settings/namedCollection.ts. The two stores were
  a literal copy-paste clone (identical CRUD facade, identical slug fn, identical
  Store interface); each now collapses to ~15 lines over one shared factory.
  Their two test files' CRUD assertions fold into one parameterized suite.
- canvas_overlay: extract the statistics engine into src/features/labMeters.ts.
  The node owned the OnlineNormalizer lifecycle + formula compilation + the
  per-frame feature sweep -- a stats engine inside a RENDERER. It now holds one
  opaque handle (mirroring what the tagging track did with taglog/presentation),
  and the lab stats are unit-testable without a canvas. The OverlayElement /
  OVERLAY_ELEMENTS / layoutCues seam is untouched -- it held across all three
  tracks that appended to it.
- dead code: delete AiDjSettingsPanel.tsx (verified unreachable even from the
  frozen legacy tree -- aiDjPlugin.activate() returns no SettingsPanel), the
  vestigial store.recordingFormats (+ its action, partialize entry, test), and
  the two never-consumed defineCollection() calls.
- SSOT: add src/util/ids.ts (slugId); four slugify clones now delegate to it,
  preserving their exact fallbacks. Ten local clamp01 re-declarations collapse
  onto the one in src/features/math.ts (the audit found five; there were ten).
- dials: split the 931-line DialsControlsPanel into primitives.tsx, labels.ts and
  panels/{voice,overlay,chord,expression,face,hand}.tsx along the seams already
  in the file. The panel is now the ~40-line composition root it should be.
- deps: drop @zodal/ui (unused), @zodal/core (unused once defineCollection went)
  and autoprefixer (no postcss config; Tailwind v4 goes through @tailwindcss/vite).
- tests: share makeRecordingCanvas (test/helpers/canvas.ts, the superset of the
  two copies) and the fixtures path + loader (test/helpers/fixtures.ts, was
  copy-pasted across six files).

Legacy engine: FROZEN, not retired (maintainer decision). src/App.tsx and
Theremin.tsx now say so explicitly: reachable only at ?engine=legacy, not being
developed, excluded from refactors by convention. Its TF/mediapipe-hands deps stay,
and src/components/, src/hooks/ and src/plugins/ai-dj/ were left out of the
dedupe/split work above.

Gates: typecheck (strict) clean, 730 tests pass, build ok, catalog regenerates
nothing (node params unchanged).
