+++ package.json
@@ -29,12 +29,10 @@
     "@tensorflow/tfjs-converter": "^4.22.0",
     "@tensorflow/tfjs-core": "^4.22.0",
     "@vitejs/plugin-react": "^5.0.4",
-    "@zodal/core": "^0.1.2",
     "@zodal/dials-core": "^0.1.0",
     "@zodal/dials-ui": "^0.1.0",
     "@zodal/store": "^0.1.2",
     "@zodal/store-localstorage": "^0.1.0",
-    "@zodal/ui": "^0.1.2",
     "acture": "^1.3.0",
     "acture-ai-vercel": "^1.1.0",
     "acture-forms-autoform": "^1.0.1",
@@ -56,7 +54,6 @@
   },
   "devDependencies": {
     "@types/node": "^22.14.0",
-    "autoprefixer": "^10.4.21",
     "tailwindcss": "^4.1.14",
     "tsx": "^4.21.0",
     "typescript": "~5.8.2",
