+++ test/settings_presets.test.ts
@@ -1,12 +1,15 @@
 /**
- * Tests the preset persistence core (src/settings) against the zodal
- * `DataProvider` contract, using an in-memory provider so it runs in Node with no
- * localStorage. Proves the backend is swappable (the same PresetStore works over
- * any provider) and that the schema round-trips voices + overlay element config.
+ * Preset-SPECIFIC persistence: what the settings schema round-trips (voices, the octave
+ * range, the overlay element config) and how it migrates/heals older preset blobs.
+ *
+ * The generic CRUD contract these presets share with the saved lab views (slug ids,
+ * save/overwrite, ordering, remove, missing → null) is asserted once for both in
+ * `named_collection.test.ts`. Uses an in-memory provider, so it runs in Node with no
+ * localStorage.
  */
 import { describe, it, expect } from 'vitest';
 import { createInMemoryProvider } from '@zodal/store';
-import { createPresetStore, presetId } from '@/settings/presets';
+import { createPresetStore } from '@/settings/presets';
 import { PresetSchema, SettingsSchema, DEFAULT_FACE_CHORD, type Preset, type Settings } from '@/settings/schema';
 import { expressionThresholds, DEFAULT_EXPRESSION_TO_DEGREE } from '@/music/expression';
 
@@ -26,19 +29,10 @@ function store() {
 }
 
 describe('preset persistence', () => {
-  it('presetId slugifies names into stable ids', () => {
-    expect(presetId('My Cool Setup!')).toBe('my-cool-setup');
-    expect(presetId('  Bossa   Nova  ')).toBe('bossa-nova');
-    expect(presetId('***')).toBe('preset'); // never empty
-  });
 
-  it('saves, lists, and round-trips a preset', async () => {
+  it('round-trips the settings payload (voices + master volume)', async () => {
     const s = store();
     await s.save('Swing Lead', sampleSettings({ masterVolume: 0.7 }), 1000);
-    const list = await s.list();
-    expect(list).toHaveLength(1);
-    expect(list[0]).toMatchObject({ id: 'swing-lead', name: 'Swing Lead', createdAt: 1000 });
-
     const loaded = await s.load('swing-lead');
     expect(loaded?.settings.masterVolume).toBeCloseTo(0.7);
     expect(loaded?.settings.right.sound).toBe('warmPad');
@@ -72,30 +66,6 @@ describe('preset persistence', () => {
     expect(loaded?.settings.left.rangeLow).toBeUndefined();
   });
 
-  it('saving the same name overwrites instead of duplicating', async () => {
-    const s = store();
-    await s.save('My Setup', sampleSettings({ masterVolume: 0.2 }), 1000);
-    await s.save('My Setup', sampleSettings({ masterVolume: 0.9 }), 2000);
-    const list = await s.list();
-    expect(list).toHaveLength(1);
-    const loaded = await s.load('my-setup');
-    expect(loaded?.settings.masterVolume).toBeCloseTo(0.9);
-    expect(loaded?.createdAt).toBe(2000);
-  });
-
-  it('orders newest first and removes by id', async () => {
-    const s = store();
-    await s.save('Old', sampleSettings(), 1000);
-    await s.save('New', sampleSettings(), 3000);
-    expect((await s.list()).map((p) => p.id)).toEqual(['new', 'old']);
-
-    await s.remove('old');
-    expect((await s.list()).map((p) => p.id)).toEqual(['new']);
-  });
-
-  it('load of a missing preset returns null', async () => {
-    expect(await store().load('nope')).toBeNull();
-  });
 
   it('rejects out-of-range settings at the schema boundary', () => {
     expect(() => sampleSettings({ masterVolume: 5 })).toThrow();
